Make edit_message_text return ApiError::MessageNotModified
Editing message text to the same value is an easy mistake to make, so I
would like to test my code against it.

I would also like to receive the error properly as an ApiError enum
variant, so this commit introduces a way for teloxide_tests actions
to return standard error values of Telegram Bot API that teloxide
can actually parse.

In the future, other usages of check_if_message_exists and ErrorBadRequest
should also be replaced with ApiErrors where possible.